+++ src/config/modelProviders/taichu.ts
@@ -4,13 +4,12 @@ import { ModelProviderCard } from '@/types/llm';
 const Taichu: ModelProviderCard = {
   chatModels: [
     {
-      description:
-        'Taichu 2.0 基于海量高质数据训练，具有更强的文本理解、内容创作、对话问答等能力',
+      contextWindowTokens: 32_768,
+      description: 'Taichu 2.0 基于海量高质数据训练，具有更强的文本理解、内容创作、对话问答等能力',
       displayName: 'Taichu 2.0',
       enabled: true,
       functionCall: true,
       id: 'taichu_llm',
-      tokens: 32_768,
     },
     /*
     // TODO: Not support for now