+++ src/server/routers/lambda/_template.ts
@@ -1,8 +1,8 @@
 import { z } from 'zod';
 
+import { insertSessionGroupSchema } from '@/database/schemas';
 import { serverDB } from '@/database/server';
 import { SessionGroupModel } from '@/database/server/models/sessionGroup';
-import { insertSessionGroupSchema } from '@/database/schemas';
 import { authedProcedure, router } from '@/libs/trpc';
 import { SessionGroupItem } from '@/types/session';