+++ src/services/message/client.ts
@@ -1,119 +1,140 @@
 import dayjs from 'dayjs';
 
-import { FileModel } from '@/database/_deprecated/models/file';
-import { MessageModel } from '@/database/_deprecated/models/message';
-import { DB_Message } from '@/database/_deprecated/schemas/message';
-import {
-  ChatFileItem,
-  ChatMessage,
-  ChatMessageError,
-  ChatTTS,
-  ChatTranslate,
-  CreateMessageParams,
-} from '@/types/message';
+import { INBOX_SESSION_ID } from '@/const/session';
+import { clientDB } from '@/database/client/db';
+import { MessageModel } from '@/database/server/models/message';
+import { BaseClientService } from '@/services/baseClientService';
+import { clientS3Storage } from '@/services/file/ClientS3';
+import { ChatMessage } from '@/types/message';
 
 import { IMessageService } from './type';
 
-export class ClientService implements IMessageService {
-  async createMessage(data: CreateMessageParams) {
-    const { id } = await MessageModel.create(data);
+export class ClientService extends BaseClientService implements IMessageService {
+  private get messageModel(): MessageModel {
+    return new MessageModel(clientDB as any, this.userId);
+  }
+
+  createMessage: IMessageService['createMessage'] = async ({ sessionId, ...params }) => {
+    const { id } = await this.messageModel.create({
+      ...params,
+      sessionId: this.toDbSessionId(sessionId) as string,
+    });
 
     return id;
-  }
+  };
 
-  async batchCreateMessages(messages: ChatMessage[]) {
-    return MessageModel.batchCreate(messages);
-  }
+  batchCreateMessages: IMessageService['batchCreateMessages'] = async (messages) => {
+    return this.messageModel.batchCreate(messages);
+  };
 
-  async getMessages(sessionId: string, topicId?: string): Promise<ChatMessage[]> {
-    const messages = await MessageModel.query({ sessionId, topicId });
+  getMessages: IMessageService['getMessages'] = async (sessionId, topicId) => {
+    const data = await this.messageModel.query(
+      {
+        sessionId: this.toDbSessionId(sessionId),
+        topicId,
+      },
+      {
+        postProcessUrl: async (url, file) => {
+          const hash = (url as string).replace('client-s3://', '');
+          const base64 = await this.getBase64ByFileHash(hash);
 
-    const fileList = (await Promise.all(
-      messages
-        .flatMap((item) => item.files)
-        .filter(Boolean)
-        .map(async (id) => FileModel.findById(id!)),
-    )) as ChatFileItem[];
+          return `data:${file.fileType};base64,${base64}`;
+        },
+      },
+    );
 
-    return messages.map((item) => ({
-      ...item,
-      imageList: fileList
-        .filter((file) => item.files?.includes(file.id) && file.fileType.startsWith('image'))
-        .map((file) => ({
-          alt: file.name,
-          id: file.id,
-          url: file.url,
-        })),
-    }));
-  }
+    return data as unknown as ChatMessage[];
+  };
 
-  async getAllMessages() {
-    return MessageModel.queryAll();
-  }
+  getAllMessages: IMessageService['getAllMessages'] = async () => {
+    const data = await this.messageModel.queryAll();
 
-  async countMessages() {
-    return MessageModel.count();
-  }
+    return data as unknown as ChatMessage[];
+  };
 
-  async countTodayMessages() {
-    const topics = await MessageModel.queryAll();
+  countMessages: IMessageService['countMessages'] = async () => {
+    return this.messageModel.count();
+  };
+
+  countTodayMessages: IMessageService['countTodayMessages'] = async () => {
+    const topics = await this.messageModel.queryAll();
     return topics.filter(
       (item) => dayjs(item.createdAt).format('YYYY-MM-DD') === dayjs().format('YYYY-MM-DD'),
     ).length;
-  }
+  };
 
-  async getAllMessagesInSession(sessionId: string) {
-    return MessageModel.queryBySessionId(sessionId);
-  }
+  getAllMessagesInSession: IMessageService['getAllMessagesInSession'] = async (sessionId) => {
+    const data = this.messageModel.queryBySessionId(this.toDbSessionId(sessionId));
 
-  async updateMessageError(id: string, error: ChatMessageError) {
-    return MessageModel.update(id, { error });
-  }
+    return data as unknown as ChatMessage[];
+  };
 
-  async updateMessage(id: string, message: Partial<DB_Message>) {
-    return MessageModel.update(id, message);
-  }
+  updateMessageError: IMessageService['updateMessageError'] = async (id, error) => {
+    return this.messageModel.update(id, { error });
+  };
 
-  async updateMessageTTS(id: string, tts: Partial<ChatTTS> | false) {
-    return MessageModel.update(id, { tts });
-  }
+  updateMessage: IMessageService['updateMessage'] = async (id, message) => {
+    return this.messageModel.update(id, message);
+  };
 
-  async updateMessageTranslate(id: string, translate: Partial<ChatTranslate> | false) {
-    return MessageModel.update(id, { translate });
-  }
+  updateMessageTTS: IMessageService['updateMessageTTS'] = async (id, tts) => {
+    return this.messageModel.updateTTS(id, tts as any);
+  };
 
-  async updateMessagePluginState(id: string, value: Record<string, any>) {
-    return MessageModel.updatePluginState(id, value);
-  }
+  updateMessageTranslate: IMessageService['updateMessageTranslate'] = async (id, translate) => {
+    return this.messageModel.updateTranslate(id, translate as any);
+  };
 
-  async updateMessagePluginArguments(id: string, value: string | Record<string, any>) {
+  updateMessagePluginState: IMessageService['updateMessagePluginState'] = async (id, value) => {
+    return this.messageModel.updatePluginState(id, value);
+  };
+
+  updateMessagePluginArguments: IMessageService['updateMessagePluginArguments'] = async (
+    id,
+    value,
+  ) => {
     const args = typeof value === 'string' ? value : JSON.stringify(value);
 
-    return MessageModel.updatePlugin(id, { arguments: args });
-  }
+    return this.messageModel.updateMessagePlugin(id, { arguments: args });
+  };
 
-  async bindMessagesToTopic(topicId: string, messageIds: string[]) {
-    return MessageModel.batchUpdate(messageIds, { topicId });
-  }
+  removeMessage: IMessageService['removeMessage'] = async (id) => {
+    return this.messageModel.deleteMessage(id);
+  };
 
-  async removeMessage(id: string) {
-    return MessageModel.delete(id);
-  }
+  removeMessages: IMessageService['removeMessages'] = async (ids) => {
+    return this.messageModel.deleteMessages(ids);
+  };
 
-  async removeMessages(ids: string[]) {
-    return MessageModel.bulkDelete(ids);
-  }
+  removeMessagesByAssistant: IMessageService['removeMessagesByAssistant'] = async (
+    sessionId,
+    topicId,
+  ) => {
+    return this.messageModel.deleteMessagesBySession(this.toDbSessionId(sessionId), topicId);
+  };
 
-  async removeMessagesByAssistant(assistantId: string, topicId?: string) {
-    return MessageModel.batchDelete(assistantId, topicId);
-  }
+  removeAllMessages: IMessageService['removeAllMessages'] = async () => {
+    return this.messageModel.deleteAllMessages();
+  };
 
-  async removeAllMessages() {
-    return MessageModel.clearTable();
-  }
-
-  async hasMessages() {
+  hasMessages: IMessageService['hasMessages'] = async () => {
     const number = await this.countMessages();
     return number > 0;
-  }
+  };
+
+  messageCountToCheckTrace: IMessageService['messageCountToCheckTrace'] = async () => {
+    const number = await this.countMessages();
+    return number >= 4;
+  };
+
+  private toDbSessionId = (sessionId: string | undefined) => {
+    return sessionId === INBOX_SESSION_ID ? undefined : sessionId;
+  };
+
+  private getBase64ByFileHash = async (hash: string) => {
+    const fileItem = await clientS3Storage.getObject(hash);
+    if (!fileItem) throw new Error('file not found');
+
+    return Buffer.from(await fileItem.arrayBuffer()).toString('base64');
+  };
 }