+++ src/store/chat/slices/builtinTool/action.test.ts
@@ -2,7 +2,7 @@ import { act, renderHook } from '@testing-library/react';
 import { describe, expect, it, vi } from 'vitest';
 
 import { fileService } from '@/services/file';
-import { ClientService } from '@/services/file/client';
+import { ClientService } from '@/services/file/_deprecated';
 import { messageService } from '@/services/message';
 import { imageGenerationService } from '@/services/textToImage';
 import { uploadService } from '@/services/upload';
@@ -47,10 +47,9 @@ describe('chatToolSlice', () => {
         url: '',
       });
       vi.spyOn(result.current, 'toggleDallEImageLoading');
-      vi.spyOn(ClientService.prototype, 'checkFileHash').mockImplementation(async () => ({
-        isExist: false,
-        metadata: {},
-      }));
+      vi.spyOn(ClientService.prototype, 'checkFileHash').mockImplementation(
+        async () => ({ isExist: false }) as any,
+      );
 
       await act(async () => {
         await result.current.generateImageFromPrompts(prompts, messageId);