+++ src/store/global/initialState.ts
@@ -1,5 +1,6 @@
 import { AppRouterInstance } from 'next/dist/shared/lib/app-router-context.shared-runtime';
 
+import { DatabaseLoadingState } from '@/types/clientDB';
 import { SessionDefaultGroup } from '@/types/session';
 import { AsyncLocalStorage } from '@/utils/localStorage';
 
@@ -37,6 +38,11 @@ export interface SystemStatus {
   hidePWAInstaller?: boolean;
   hideThreadLimitAlert?: boolean;
   inputHeight: number;
+  /**
+   * 应用初始化时不启用 PGLite，只有当用户手动开启时才启用
+   */
+  isEnablePglite?: boolean;
+  latestChangelogId?: string;
   mobileShowPortal?: boolean;
   mobileShowTopic?: boolean;
   sessionsWidth: number;
@@ -50,6 +56,13 @@ export interface SystemStatus {
 
 export interface GlobalState {
   hasNewVersion?: boolean;
+  initClientDBError?: Error;
+  initClientDBProcess?: { costTime?: number; phase: 'wasm' | 'dependencies'; progress: number };
+  /**
+   * 客户端数据库初始化状态
+   * 启动时为 Idle，完成为 Ready，报错为 Error
+   */
+  initClientDBStage: DatabaseLoadingState;
   isMobile?: boolean;
   isStatusInit?: boolean;
   latestVersion?: string;
@@ -76,6 +89,7 @@ export const INITIAL_STATUS = {
 } satisfies SystemStatus;
 
 export const initialState: GlobalState = {
+  initClientDBStage: DatabaseLoadingState.Idle,
   isMobile: false,
   isStatusInit: false,
   sidebarKey: SidebarTabKey.Chat,